refactor components & create UploadModal (#180)
* refactor React components into own files

This will make each component easier to maintain and allows a more
modular design.

This is done with: ObjectsList, SideBar, Path, BrowserUpdate, and
ConfirmModal, all pre-existing components.

* create UploadModal for upload progress

This replaces code previously in the Browse component to create a
modular, React component replacement. It deals with the abort modal and
the upload progress bar at once.