fix state updates and imports (#185)
This fixes a bug that I introduced (ah ha ha... :L) where the state
would not update because redux only shallowly compares objects in the
state to check whether or not to update components that subscribe to
those changes. Because I only modified the object beneath rather than
modifying the object at the top level, it wouldn't usually trigger
updates, although it would trigger updates every so often (for reasons I
don't know).